update trigger and delete processed items from cursor
